--- client/src/components/Footer/LogoMedia.js
@@ -1,133 +1,0 @@
-import React from "react";
-import { NavLink } from "react-router-dom";
-import Logo from "../../helpers/images/logo.png";
-
-const Media = () => {
-  return (
-    <div className="col l2 s12">
-      <div className="nav-wrapper cyan darken-1">
-        <NavLink to="/" exact>
-          <img
-            src={Logo}
-            style={{
-              display: "inline-block",
-              height: "62px",
-            }}
-          />
-          <NavLink
-            to="/"
-            exact
-            style={{
-              marginLeft: "25px",
-              position: "absolute",
-              color: "#fff",
-              display: "inline-block",
-              fontSize: "2.1rem",
-              padding: "0",
-            }}
-          >
-            TUTORS
-          </NavLink>
-        </NavLink>
-      </div>
-      <section>
-        <h2 hidden>social media</h2>
-        <h5>Follow us:</h5>
-        <ul
-          style={{
-            display: "flex",
-            justifyContent: "space-between",
-            paddingTop: "10px",
-          }}
-        >
-          <li>
-            <a href="#" aria-label="facebook">
-              <svg
-                version="1.1"
-                xmlns="http://www.w3.org/2000/svg"
-                width="32"
-                height="32"
-                viewBox="0 0 32 32"
-              >
-                <title>facebook</title>
-                <path
-                  fill="#fff"
-                  d="M21.329 5.313h2.921v-5.088c-0.504-0.069-2.237-0.225-4.256-0.225-4.212 0-7.097 2.649-7.097 7.519v4.481h-4.648v5.688h4.648v14.312h5.699v-14.311h4.46l0.708-5.688h-5.169v-3.919c0.001-1.644 0.444-2.769 2.735-2.769z"
-                ></path>
-              </svg>
-            </a>
-          </li>
-          <li>
-            <a href="#" aria-label="twitter">
-              <svg
-                version="1.1"
-                xmlns="http://www.w3.org/2000/svg"
-                width="32"
-                height="32"
-                viewBox="0 0 32 32"
-              >
-                <title>twitter</title>
-                <path
-                  fill="#fff"
-                  d="M32 6.078c-1.19 0.522-2.458 0.868-3.78 1.036 1.36-0.812 2.398-2.088 2.886-3.626-1.268 0.756-2.668 1.29-4.16 1.588-1.204-1.282-2.92-2.076-4.792-2.076-3.632 0-6.556 2.948-6.556 6.562 0 0.52 0.044 1.020 0.152 1.496-5.454-0.266-10.28-2.88-13.522-6.862-0.566 0.982-0.898 2.106-0.898 3.316 0 2.272 1.17 4.286 2.914 5.452-1.054-0.020-2.088-0.326-2.964-0.808 0 0.020 0 0.046 0 0.072 0 3.188 2.274 5.836 5.256 6.446-0.534 0.146-1.116 0.216-1.72 0.216-0.42 0-0.844-0.024-1.242-0.112 0.85 2.598 3.262 4.508 6.13 4.57-2.232 1.746-5.066 2.798-8.134 2.798-0.538 0-1.054-0.024-1.57-0.090 2.906 1.874 6.35 2.944 10.064 2.944 12.072 0 18.672-10 18.672-18.668 0-0.29-0.010-0.57-0.024-0.848 1.302-0.924 2.396-2.078 3.288-3.406z"
-                ></path>
-              </svg>
-            </a>
-          </li>
-          <li>
-            <a href="#" aria-label="instagram">
-              <svg
-                version="1.1"
-                xmlns="http://www.w3.org/2000/svg"
-                width="32"
-                height="32"
-                viewBox="0 0 32 32"
-              >
-                <title>instagram</title>
-                <path
-                  fill="#fff"
-                  d="M31.969 9.408c-0.075-1.7-0.35-2.869-0.744-3.882-0.406-1.075-1.031-2.038-1.85-2.838-0.8-0.813-1.769-1.444-2.832-1.844-1.019-0.394-2.182-0.669-3.882-0.744-1.713-0.081-2.257-0.1-6.601-0.1s-4.888 0.019-6.595 0.094c-1.7 0.075-2.869 0.35-3.882 0.744-1.075 0.406-2.038 1.031-2.838 1.85-0.813 0.8-1.444 1.769-1.844 2.832-0.394 1.019-0.669 2.182-0.744 3.882-0.081 1.713-0.1 2.257-0.1 6.601s0.019 4.888 0.094 6.595c0.075 1.7 0.35 2.869 0.744 3.882 0.406 1.075 1.038 2.038 1.85 2.838 0.8 0.813 1.769 1.444 2.832 1.844 1.019 0.394 2.182 0.669 3.882 0.744 1.706 0.075 2.25 0.094 6.595 0.094s4.888-0.019 6.595-0.094c1.7-0.075 2.869-0.35 3.882-0.744 2.151-0.832 3.851-2.532 4.682-4.682 0.394-1.019 0.669-2.182 0.744-3.882 0.075-1.707 0.094-2.251 0.094-6.595s-0.006-4.888-0.081-6.595zM29.087 22.473c-0.069 1.563-0.331 2.407-0.55 2.969-0.538 1.394-1.644 2.5-3.038 3.038-0.563 0.219-1.413 0.481-2.969 0.55-1.688 0.075-2.194 0.094-6.464 0.094s-4.782-0.019-6.464-0.094c-1.563-0.069-2.407-0.331-2.969-0.55-0.694-0.256-1.325-0.663-1.838-1.194-0.531-0.519-0.938-1.144-1.194-1.838-0.219-0.563-0.481-1.413-0.55-2.969-0.075-1.688-0.094-2.194-0.094-6.464s0.019-4.782 0.094-6.464c0.069-1.563 0.331-2.407 0.55-2.969 0.256-0.694 0.663-1.325 1.2-1.838 0.519-0.531 1.144-0.938 1.838-1.194 0.563-0.219 1.413-0.481 2.969-0.55 1.688-0.075 2.194-0.094 6.464-0.094 4.276 0 4.782 0.019 6.464 0.094 1.563 0.069 2.407 0.331 2.969 0.55 0.694 0.256 1.325 0.662 1.838 1.194 0.531 0.519 0.938 1.144 1.194 1.838 0.219 0.563 0.481 1.413 0.55 2.969 0.075 1.688 0.094 2.194 0.094 6.464s-0.019 4.77-0.094 6.458z"
-                ></path>
-                <path
-                  fill="#fff"
-                  d="M16.059 7.783c-4.538 0-8.22 3.682-8.22 8.22s3.682 8.22 8.22 8.22c4.539 0 8.22-3.682 8.22-8.22s-3.682-8.22-8.22-8.22zM16.059 21.335c-2.944 0-5.332-2.388-5.332-5.332s2.388-5.332 5.332-5.332c2.944 0 5.332 2.388 5.332 5.332s-2.388 5.332-5.332 5.332z"
-                ></path>
-                <path
-                  fill="#fff"
-                  d="M26.524 7.458c0 1.060-0.859 1.919-1.919 1.919s-1.919-0.859-1.919-1.919c0-1.060 0.859-1.919 1.919-1.919s1.919 0.859 1.919 1.919z"
-                ></path>
-              </svg>
-            </a>
-          </li>
-          <li>
-            <a href="#" aria-label="linkedin">
-              <svg
-                version="1.1"
-                xmlns="http://www.w3.org/2000/svg"
-                width="32"
-                height="32"
-                viewBox="0 0 32 32"
-              >
-                <title>linkedin</title>
-                <path
-                  fill="#fff"
-                  d="M31.992 32v-0.001h0.008v-11.736c0-5.741-1.236-10.164-7.948-10.164-3.227 0-5.392 1.771-6.276 3.449h-0.093v-2.913h-6.364v21.364h6.627v-10.579c0-2.785 0.528-5.479 3.977-5.479 3.399 0 3.449 3.179 3.449 5.657v10.401h6.62z"
-                ></path>
-                <path
-                  fill="#fff"
-                  d="M0.528 10.636h6.635v21.364h-6.635v-21.364z"
-                ></path>
-                <path
-                  fill="#fff"
-                  d="M3.843 0c-2.121 0-3.843 1.721-3.843 3.843s1.721 3.879 3.843 3.879c2.121 0 3.843-1.757 3.843-3.879-0.001-2.121-1.723-3.843-3.843-3.843v0z"
-                ></path>
-              </svg>
-            </a>
-          </li>
-        </ul>
-      </section>
-    </div>
-  );
-};
-
-export default Media;
