+++ client/src/index.css
@@ -25,8 +25,29 @@ nav,
   box-shadow: none;
 }
 
-svg:hover {
-  padding: 7px;
-  border-radius: 30%;
-  background-color: #00838f;
+a.social {
+  padding-right: 20px;
+  transition: transform 250ms;
+  display: inline-block;
+}
+
+a.social:hover {
+  transform: translateY(-2px);
+}
+a.youtube {
+  padding-left: 5px;
+  color: #eb3223;
+}
+
+a.facebook {
+  color: #4968ad;
+}
+
+a.twitter {
+  color: #3581be;
+}
+
+a.instagram {
+  color: rgba(128, 0, 128, 0.541);
+  padding-right: 0px;
 }
